+++ common/src/main/java/net/more_rpg_classes/effect/IgnitedEffect.java
@@ -9,7 +9,6 @@ public class IgnitedEffect extends CustomStatusEffect {
         super(category, color);
     }
 
-
     @Override
     public boolean canApplyUpdateEffect(int duration, int amplifier) {
         return duration % 10 == 0;